fix(mirror): satisfy clippy and 1k-line gate for LAN bind
Drop the unused MirrorHost::allow_lan getter, cfg(test) pick/rank helpers
that clippy -D warnings treated as dead on the lib target, and move the
LAN bind integration test out of mirror/mod.rs so the file stays under
the FILES_OVER_1K_BUDGET ceiling.

Refs #875
